+++ src/consumer/component/consumer-component.js
@@ -404,8 +404,10 @@ export default class Component {
       return { name: file.basename, relativePath: pathNormalizeToLinux(file.relative), test: file.test };
     });
 
+    const componentId = !this.scope && this.version ? this.id.cloneNoVersion() : this.id;
+
     return bitMap.addComponent({
-      componentId: this.id,
+      componentId,
       files: filesForBitMap,
       mainFile: this.mainFile,
       rootDir,